% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/swag.r
\name{swag}
\alias{swag}
\title{Build an R web API package from a swagger definition}
\usage{
swag(spec, pkg_name = NULL, target_dir = "/Users/bob/test",
  overwrite = FALSE)
}
\arguments{
\item{spec}{URL or path to swagger JSON specification}

\item{pkg_name}{name for the package. if \code{NULL} it will be derived
from the \code{"title"} field in the \code{info} section of the
swagger description.}

\item{target_dir}{where to put the package. a temp dir will be used if none
is provided}

\item{overwrite}{should we overwrite \code{target_dir} if it exists? (default: \code{FALSE})}
}
\value{
path to the newly created package path
}
\description{
Build an R web API package from a swagger definition
}

