% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/swagger-package.r
\docType{package}
\name{swagger}
\alias{swagger}
\alias{swagger-package}
\title{Build R API Packages Automatically from Swagger <http://swagger.io> JSON Specifications}
\description{
Build R API Packages Automatically from Swagger <http://swagger.io> JSON Specifications
}
\author{
Bob Rudis (@hrbrmstr)
}

