% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{swagger_spec}
\alias{swagger_spec}
\title{Swagger Index File with OpenAPI Path}
\usage{
swagger_spec(apiPath = "http://petstore.swagger.io/v2/swagger.json")
}
\arguments{
\item{apiPath}{Path to an openAPI specification}
}
\value{
large string containing the contents of \code{\link{swagger_index}()} with
  the appropriate speicification path changed to the \code{apiPath} value.
}
\description{
Produces a \code{index.html} file that will attempt to access a provided API path.
}
\examples{
if (interactive()) {
  swagger_spec("https://developer.uspto.gov/ibd-api/swagger.json")
}
}
