% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swagger.R
\docType{package}
\name{swagger-package}
\alias{swagger}
\alias{swagger-package}
\title{swagger: Dynamically Generates Documentation from a Swagger-Compliant API}
\description{
Swagger is a collection of HTML, Javascript, and CSS assets that
dynamically generate beautiful documentation from a Swagger-compliant API.
}
\details{
The main purpose of this package is to enable package authors to create
APIs that are compatible with \url{https://swagger.io/}
and \url{https://www.openapis.org/}.

To learn more about Swagger visit: \url{https://swagger.io/swagger-ui/}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rstudio/swagger}
  \item Report bugs at \url{https://github.com/rstudio/swagger/issues}
}

}
\author{
\strong{Maintainer}: Javier Luraschi \email{javier@rstudio.com}

Authors:
\itemize{
  \item  SmartBear Software [copyright holder]
}

Other contributors:
\itemize{
  \item  RStudio [copyright holder]
}

}
