% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swagger.R
\docType{package}
\name{swagger-package}
\alias{swagger}
\alias{swagger-package}
\title{swagger: Dynamically Generates Documentation from a Swagger-Compliant API}
\description{
Swagger is a collection of HTML, JavaScript, and CSS assets that
dynamically generate beautiful documentation from a Swagger-compliant API.
}
\details{
The main purpose of this package is to enable package authors to create
APIs that are compatible with \url{https://swagger.io/}
and \url{https://www.openapis.org/}.

To learn more about Swagger visit: \url{https://swagger.io/swagger-ui/}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rstudio.github.io/swagger/}
  \item \url{https://github.com/rstudio/swagger}
  \item Report bugs at \url{https://github.com/rstudio/swagger/issues}
}

}
\author{
\strong{Maintainer}: Bruno Tremblay \email{cran@neoxone.com} [contributor]

Authors:
\itemize{
  \item Barret Schloerke \email{barret@rstudio.com} (\href{https://orcid.org/0000-0001-9986-114X}{ORCID})
  \item Javier Luraschi \email{javier@rstudio.com}
  \item  SmartBear Software [copyright holder]
}

Other contributors:
\itemize{
  \item  RStudio [copyright holder]
}

}
