% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{swagger_index}
\alias{swagger_index}
\title{Path to Swagger Index}
\usage{
swagger_index()
}
\description{
Retrieves the path to the swagger index file.
}
\examples{

if (interactive()) {
  browseURL(swagger_index())
} else {
  print(paste("You can use swagger under: ", swagger_index()))
}

}
