% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{swagger_path}
\alias{swagger_path}
\title{Path to Swagger Resources}
\usage{
swagger_path()
}
\description{
Retrieves the path to swagger resources.
}
\examples{

if (interactive()) {
  browseURL(swagger_path())
} else {
  print(paste("You can explore swagger resources under: ", swagger_path()))
}

}
