% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{swagger_path}
\alias{swagger_path}
\title{Path to Swagger Resources}
\usage{
swagger_path(version = "3")
}
\arguments{
\item{version}{Major version number to be used for Swagger UI.  Currently version \code{"3"} is compatible with OpenAPI spec 2.0, 3.0.}
}
\description{
Retrieves the path to swagger resources.
}
\examples{

if (interactive()) {
  browseURL(swagger_path())
} else {
  print(paste("You can explore swagger resources under: ", swagger_path()))
}

}
