% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{swagger_spec}
\alias{swagger_spec}
\title{Swagger Index File with OpenAPI Path}
\usage{
swagger_spec(
  api_path = "\\"http://petstore.swagger.io/v2/swagger.json\\"",
  version = "5"
)
}
\arguments{
\item{api_path}{Path to paste into the an OpenAPI specification file}

\item{version}{Major version number to be used for Swagger UI.  Currently version \code{"3"} is compatible with OpenAPI spec 2.0, 3.0.}
}
\value{
large string containing the contents of \code{\link{swagger_index}()} with
  the appropriate specification path changed to the \code{api_path} value.
}
\description{
Produces the content for a \code{index.html} file that will attempt
to access a provided API path.
}
\examples{
if (interactive()) {
  swagger_spec("https://developer.uspto.gov/ibd-api/swagger.json")
}
}
